pyproject.toml
[project]
name = "mcp-clickhouse"
version = "0.1.12"
description = "An MCP server for ClickHouse."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.10"
dependencies = [
     "fastmcp>=2.0.0",
     "python-dotenv>=1.0.1",
     "clickhouse-connect>=0.8.16",
     "truststore>=0.10",
     "chdb>=3.3.0",
     "psycopg2-binary>=2.9.9",
     "pgvector>=0.2.5",
]

[project.scripts]
mcp-clickhouse = "mcp_server.main:main"

[project.urls]
Home = "https://github.com/OriginHubAI/mcp-sqlvectordb"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio"
]

[tool.hatch.build.targets.wheel]
packages = ["mcp_server"]

[tool.ruff]
line-length = 100

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
